requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
elastic-transport==8.15.0
elasticsearch==8.15.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.2
fastapi-cli==0.0.5
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.10
jax==0.4.33
jaxlib==0.4.33
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.14
langchain-community==0.2.15
langchain-core==0.2.35
langchain-text-splitters==0.2.2
langsmith==0.1.106
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
more-itertools==10.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai==1.44.1
openai-whisper==20231117
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.7
packaging==24.1
pillow==10.4.0
pip==24.2
protobuf==4.25.4
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
pypdf==4.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
safetensors==0.4.4
scipy==1.14.1
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
soupsieve==2.5
SQLAlchemy==2.0.32
starlette==0.38.2
sympy==1.13.2
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typer-slim==0.12.5
typing_extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.23.0
websockets==13.0.1
wheel==0.44.0
win-inet-pton==1.1.0
yarl==1.11.1
zstandard==0.23.0
